Setting up YUIdocs configuration, and some examples.
Created ‘yuidoc.json’ to set the configuration for auto-generation of
the app’s documentation.
Edited .gitignore to ignore the YUIdocs directory that will be created
by YUIdocs; no version control is required for the output directory.
Added comments that are recognizable by YUIdocs, as first example of
how to write them and to show some results.